Débogage en U3 : gestion des adresses 32 et 64 bits

+++ c/cretinlib/utile_32.h
@@ -1,7 +1,7 @@
 # ifndef __CRETINLIB_UTILE__
 # define __CRETINLIB_UTILE__
 
-# define TYPE_ADRESSE (unsigned long long int)
+# define TYPE_ADRESSE (unsigned long int)
 
 int adresse(void * truc);
 int adresseComplete(void * truc);
